Version estable tras la reunion del 23/05/2024

+++ app/src/main/java/com/alvdela/smartspend/model/CashFlow.kt
@@ -1,10 +1,11 @@
 package com.alvdela.smartspend.model
 
+import java.math.BigDecimal
 import java.time.LocalDate
 
 data class CashFlow (
     val description: String = "gasto",
-    val amount: Float,
+    val amount: BigDecimal,
     val type: CashFlowType,
     val date: LocalDate
 )
